+++ Net2Plan-Core/src/main/java/com/net2plan/interfaces/IGUIPluginWrapper.java
@@ -1,4 +1,4 @@
-package com.net2plan.gui.launcher.wrapper;
+package com.net2plan.interfaces;
 
 import java.util.Map;
