+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/debug/GUINetworkDesignWrapper.java
@@ -1,6 +1,7 @@
-package com.net2plan.gui.launcher.wrapper;
+package com.net2plan.gui.plugins.debug;
 
-import com.net2plan.gui.launcher.utils.GUIRobot;
+import com.net2plan.gui.utils.GUIRobot;
+import com.net2plan.interfaces.IGUIPluginWrapper;
 
 import java.awt.*;
 import java.awt.event.KeyEvent;
